+++ lib/src/models.dart
@@ -1,5 +1,7 @@
 library dart_holodex_api.models;
 
+import 'package:dart_holodex_api/src/utils/enum_util.dart';
+
 import 'enums.dart';
 import 'package:equatable/equatable.dart';
 import 'dart:convert';